240309 moved notice and qna api view to their directory
